+++ trifusion/TriFusion.py
@@ -15,6 +15,7 @@ def gui_exec():
 
     main()
 
+
 if __name__ == "__main__":
 
     import multiprocessing